pyproject.toml
[tool.poetry]
name = "simplest-flask-oauth-app"
version = "0.1.0"
description = ""
authors = ["Eduardo Ballesteros <eduardo.ballesterosg@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.0"
gunicorn = "^23.0.0"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
